requirements.txt
opencv-python
pytesseract
googletrans==4.0.0rc1
Pillow
pygetwindow
pyautogui
